requirements.txt
python-telegram-bot==21.7
openai==1.66.3
notion-client==2.2.1
python-dotenv==1.0.1
